New Class PlaceHolderTextField to create a place holder inside some JTextField extended.
News by LogicModelNewParking, NewParking, NewParkingController and more.

DAOParking lengthParkingTAbleDataBase is now a long type variable.

DaoParkingImpl removed newParking instance by constructor method. It
will be created by the addNewParkingReservation Method.

DaoParkingImpl displayListParking method addKeyListener to
newParkingReservationJButton and also to
cancelNewParkingReservationJButton.

by LogicModelNewParking we create many variables to handle with errors
JDialog, ActionListener and KeyListener was added to the OK JButtons of
each JDialog errors.

LogicModelNewParking: new checkAnreiseDateFormat,
checkAbreiseDateFormat, setReadyToSaveArrival, SetReadyToSaveDeparture,
checkAllEntries Methods. all methods have their proper descriptions and
comments.

By NewParking.java we have initialized certain objects and more. See the
Class to check the changes.

NewParkingController: implements FocusListener; we add some variables
for closing New Parking Reservation Form; Constructor receive a
LogicModelNewParking; We add some Listener.


NewParkingController actionPerformed method call logicModelNewParkirg:
setReadyToSaveArrival(true), setReadyToSaveDeparture(true),
checkAnreiseDateFormat method, checkAbreiseDAteFormat method,
checkAllEntries method etc. See for more info.

NewParkingController some new instructions by focusGained. See more.

NewParkingControlleer also some new instructions by focusLost.

+++ BigDataAirportHotelBasel/src/tech/codepalace/dao/DAOParking.java
@@ -26,7 +26,7 @@ public interface DAOParking {
 	/** Method to add a new entry in the database
 	 *  @author Antonio Estevez Gonzalez
 	 **/
-	void addNewParkingReservation(int lenghtParkingTableDataBase) throws DaoException;
+	void addNewParkingReservation(long lenghtParkingTableDataBase) throws DaoException;
